reorganize LICENSE info
