+++ src/features/effects/effectFactory.js
@@ -2,7 +2,7 @@ import FineTuning from "./FineTuning"
 import Fisheye from "./Fisheye"
 import Loop from "./Loop"
 import Mask from "./Mask"
-import Noise from "./Noise"
+import Noise from "./noise/Noise"
 import Track from "./Track"
 import Transformer from "./Transformer"
 import Warp from "./Warp"
